# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Reloading Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reference/{category}:
    get:
      operationId: listReferenceData
      tags: [reference]
      summary: List reference data by category
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reference data list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReferenceItem"
    post:
      operationId: createReferenceItem
      tags: [reference]
      summary: Add a reference data item
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReferenceItemBody"
      responses:
        "201":
          description: Created item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferenceItem"
  /reference/{category}/{id}:
    patch:
      operationId: updateReferenceItem
      tags: [reference]
      summary: Update a reference data item
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReferenceItemBody"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferenceItem"
    delete:
      operationId: deleteReferenceItem
      tags: [reference]
      summary: Delete a reference data item
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Settings ─────────────────────────────────────────────────
...
  /loads:
    get:
      operationId: listLoads
      tags: [loads]
      summary: List all loads
      responses:
        "200":
          description: List of loads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Load"
    post:
      operationId: createLoad
      tags: [loads]
      summary: Create a new load
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLoadBody"
      responses:
        "201":
          description: Created load
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Load"
  /loads/{id}:
    get:
      operationId: getLoad
      tags: [loads]
      summary: Get a load
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Load
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Load"
    patch:
      operationId: updateLoad
      tags: [loads]
      summary: Update load data / advance step
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLoadBody"
      responses:
        "200":
          description: Updated load
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Load"
    delete:
      operationId: deleteLoad
      tags: [loads]
      summary: Delete a load (with optional restock)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteLoadBody"
      responses:
        "204":
          description: Deleted
  /loads/{id}/complete:
    post:
      operationId: completeLoad
      tags: [loads]
      summary: Mark load as completed, deduct inventory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed load
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Load"
  /loads/{id}/fire:
    post:
      operationId: fireLoad
      tags: [loads]
      summary: Mark completed load as fired
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FireLoadBody"
      responses:
        "200":
          description: Fired load
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Load"

  # ── Charge Ladders ────────────────────────────────────────────
...
components:
  schemas:
...
    ReferenceItem:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
        value:
          type: string
        sortOrder:
          type: integer
      required: [id, category, value, sortOrder]
    CreateReferenceItemBody:
      type: object
      properties:
        value:
          type: string
        sortOrder:
          type: integer
      required: [value]
    UpdateReferenceItemBody:
      type: object
      properties:
        value:
          type: string
        sortOrder:
          type: integer
...
    Load:
      type: object
      properties:
        id:
          type: integer
        loadNumber:
          type: integer
          nullable: true
        userLoadId:
          type: string
          nullable: true
        cartridgeId:
          type: integer
        cartridgeProductionCharge:
          type: string
        reloadingCycle:
          type: integer
        date:
          type: string
          format: date
        caliber:
          type: string
        cartridgeQuantityUsed:
          type: integer
        primerId:
          type: integer
          nullable: true
        primerQuantityUsed:
          type: integer
          nullable: true
        powderId:
          type: integer
          nullable: true
        powderChargeGr:
          type: number
          nullable: true
        powderTotalUsedGr:
          type: number
          nullable: true
        bulletId:
          type: integer
          nullable: true
        bulletQuantityUsed:
          type: integer
          nullable: true
        coalIn:
          type: number
          nullable: true
        oalIn:
          type: number
          nullable: true
        l6In:
          type: number
          nullable: true
        washingMinutes:
          type: integer
          nullable: true
        annealingMinutes:
          type: integer
          nullable: true
        secondWashingMinutes:
          type: integer
          nullable: true
        calibrationType:
          type: string
          nullable: true
        skippedSteps:
          type: string
          nullable: true
        h2oWeightGr:
          type: number
          nullable: true
        photoBase64:
          type: string
          nullable: true
        completed:
          type: boolean
        fired:
          type: boolean
        notes:
          type: string
          nullable: true
        deletedAt:
          type: string
          format: date-time
          nullable: true
        deletedNote:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, cartridgeId, cartridgeProductionCharge, reloadingCycle, date, caliber, cartridgeQuantityUsed, completed, fired, createdAt]
    CreateLoadBody:
      type: object
      properties:
        cartridgeId:
          type: integer
        cartridgeQuantityUsed:
          type: integer
        notes:
          type: string
      required: [cartridgeId, cartridgeQuantityUsed]
    UpdateLoadBody:
      type: object
      properties:
        primerId:
          type: integer
        primerQuantityUsed:
          type: integer
        powderId:
          type: integer
        powderChargeGr:
          type: number
        bulletId:
          type: integer
        bulletQuantityUsed:
          type: integer
        coalIn:
          type: number
        oalIn:
          type: number
        l6In:
          type: number
        washingMinutes:
          type: integer
        annealingMinutes:
          type: integer
        secondWashingMinutes:
          type: integer
        calibrationType:
          type: string
        skippedSteps:
          type: string
        photoBase64:
          type: string
          nullable: true
        notes:
          type: string
    DeleteLoadBody:
      type: object
      properties:
        restockPrimers:
          type: integer
          description: Number of primers to restock
        restockPowderGr:
          type: number
          description: Grams of powder to restock
        restockBullets:
          type: integer
          description: Number of bullets to restock
        note:
          type: string
          description: Reason for deletion
    FireLoadBody:
      type: object
      properties:
        h2oWeightGr:
          type: number
...