# lib/api-zod/src/generated/types/deleteLoadBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reloading Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface DeleteLoadBody {
  /** Number of primers to restock */
  restockPrimers?: number;
  /** Grams of powder to restock */
  restockPowderGr?: number;
  /** Number of bullets to restock */
  restockBullets?: number;
  /** Reason for deletion */
  note?: string;
}