# lib/api-zod/src/generated/types/updateReferenceItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reloading Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateReferenceItemBody {
  value?: string;
  sortOrder?: number;
}